Restructured the python files
